put Metermap in its own file
